Account for more possible Fly machine states
The states previously included here are mentioned in the API spec, but I
think mentioned only as an example of how to provided a comma-separated
list of states to filter, or the valid states that can be waited for
using the /wait endopint. They also reference a table for the full list 
of possible states: https://fly.io/docs/machines/machine-states/

Notably, one of the states I've included here ('replaced') does not
appear in this table, but I've personally observed this state when
trying to deserialize with an enum that included all the other states
listed in the above table. A documentation issue was filed for that
here: https://github.com/superfly/docs/issues/1835