Align wait_for_machine_state's response with Fly Machines API
Fly API doesn't respond to /wait with any of the fields of MachineResponse,
so that will always deserialize to a struct of 'None' fields, and that 
signature is a bit confusing in the docs if the user expects to actually 
get a meaningful MachineResponse. Instead, return unit type, which more
closely matches how Fly Machines API responds to this request.